Move datagrid components to datagrid subdirectory
