+++ packages/ra-ui-materialui/src/detail/showFieldTypes.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Datagrid from '../list/Datagrid';
+import Datagrid from '../list/datagrid/Datagrid';
 import ArrayField from '../field/ArrayField';
 import BooleanField from '../field/BooleanField';
 import DateField from '../field/DateField';
