+++ packages/ra-ui-materialui/src/list/index.js
@@ -1,12 +1,5 @@
 import BulkActionsToolbar from './BulkActionsToolbar';
 import BulkDeleteAction from './BulkDeleteAction';
-import Datagrid from './Datagrid';
-import DatagridLoading from './DatagridLoading';
-import DatagridBody, { PureDatagridBody } from './DatagridBody';
-import DatagridRow, { PureDatagridRow } from './DatagridRow';
-import DatagridHeaderCell from './DatagridHeaderCell';
-import DatagridCell from './DatagridCell';
-import ExpandRowButton from './ExpandRowButton';
 import List from './List';
 import ListActions from './ListActions';
 import ListGuesser from './ListGuesser';
@@ -21,17 +14,11 @@ import SimpleListLoading from './SimpleListLoading';
 import SingleFieldList from './SingleFieldList';
 
 export * from './filter';
+export * from './datagrid';
 
 export {
     BulkActionsToolbar,
     BulkDeleteAction,
-    Datagrid,
-    DatagridLoading,
-    DatagridBody,
-    DatagridRow,
-    DatagridHeaderCell,
-    DatagridCell,
-    ExpandRowButton,
     List,
     ListView,
     ListActions,
@@ -41,8 +28,6 @@ export {
     PaginationActions,
     PaginationLimit,
     Placeholder,
-    PureDatagridBody,
-    PureDatagridRow,
     SimpleList,
     SimpleListLoading,
     SingleFieldList,
